feat : Score 값을 실수로 변경

+++ src/main/java/chess/domain/Score.java
@@ -4,9 +4,9 @@ import java.util.Objects;
 
 public class Score {
 
-    private final int value;
+    private final Double value;
 
-    Score(int value) {
+    Score(Double value) {
         this.value = value;
     }
